Simplify and document the setup
Now: just require one authorization token, not two. The path of the url changes from `/mobile` to `/public`.

Document how to get this token, and how to get the theater ids.
☝️Note that the single token obtained with these setups is associated with an allocine account. The previous system with the `/mobile` path and two tokens was anonymous.
I don't know how to obtain the anonymous token.

It appears the duration of both the anonymous and user-associated tokens is one year.

Also changed some theater ids in the examples to be consistent with the setup instructions.